==> snake/1. Data Prepare.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score

import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

train = pd.read_csv("../data/age_train.csv",names=['uid','age_group']).sort_values(by=['uid'])
test = pd.read_csv("../data/age_test.csv",names=['uid']).sort_values(by=['uid'])
info = pd.read_csv("../data/app_info.csv",names=['appid','category'])
active = pd.read_csv("../data/user_app_actived.csv",names=['uid','appid']).sort_values(by=['uid'])
usage = pd.read_csv("../data/user_app_usage.csv",names=['uid','appid','duration','times','use_date'],parse_dates=['use_date'])
user_basic_info = pd.read_csv("../data/user_basic_info.csv",names=['uid','gender','city','prodname','ramcapacity','ramleftration','romcapacity','romleftration','color','fontsize','ct','carrier','os']).sort_values(by=['uid'])
behavior_info = pd.read_csv("../data/user_behavior_info.csv",names=['uid','boottimes','a','b','c','d','e','f','g']).sort_values(by=['uid'])
print((train.shape,test.shape),(info.shape,active.shape,user_basic_info.shape,behavior_info.shape))#usage.shape,


pickle_path = "../pickle"
if not os.path.exists(pickle_path):
    os.mkdir(pickle_path)

if not os.path.exists("{}/user_app_usage.pickle".format(pickle_path)):
    t1 = time.time()
    usage.to_pickle("{}/user_app_usage.pickle".format(pickle_path))
    print('USAGE TO PICKLE: ',time.time()-t1)

usage_app_seq = usage[['uid','appid']].groupby(['uid'])['appid'].apply(lambda x:list(x)).reset_index()
usage_app_seq.to_pickle("{}/user_app_seq.pickle".format(pickle_path))

from tqdm import tqdm

def flatten_active(df):    
    u = []
    a = []
    for i in tqdm(range(len(df['appid'].values))):
        u += [df['uid'].values[i]]*df['app_len'].values[i]
        a += list(df['appid'].values[i])
        
    new_df = pd.DataFrame()
    new_df['uid'] = u
    new_df['appid'] = a
        
    return new_df

active['appid'] = active['appid'].map(lambda x:x.split('#'))
active['app_len'] = active['appid'].map(lambda x:len(x))
active = active.reset_index(drop=True)
deal_active = flatten_active(active)

if not os.path.exists("{}/user_app_active.pickle".format(pickle_path)):
    t1 = time.time()
    active.to_pickle("{}/user_app_active.pickle".format(pickle_path))
    print('ACTIVE TO PICKLE: ',time.time()-t1)

if not os.path.exists("{}/user_app_active_flatten.pickle".format(pickle_path)):
    t1 = time.time()
    deal_active.to_pickle("{}/user_app_active_flatten.pickle".format(pickle_path))
    print('Deal ACTIVE TO PICKLE: ',time.time()-t1)

active = pd.read_pickle("../pickle/user_app_active.pickle")

active.head()

all_data = train.append(test)
all_data = all_data.sort_values(by=['uid']).reset_index(drop=True)

active_train = active.merge(train,how='right',on='uid')
active_test = active.merge(test,how='right',on='uid')
active_train.to_pickle("../pickle/active_text_train.pickle")
active_test.to_pickle("../pickle/active_text_test.pickle")

usage_train = usage.merge(train,how='right',on='uid')
usage_test = usage.merge(test,how='right',on='uid')
usage_train.to_pickle("../pickle/usage_text_train.pickle")
usage_test.to_pickle("../pickle/usage_text_test.pickle")
