==> snake/4. Flatten Stat Feature.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score

import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

pickle_path = "../pickle"
active = pd.read_pickle("{}/user_app_active_flatten.pickle".format(pickle_path))
usage = pd.read_pickle("{}/user_app_usage.pickle".format(pickle_path))

train = pd.read_csv("../data/age_train.csv",names=['uid','age_group']).sort_values(by=['uid'])
test = pd.read_csv("../data/age_test.csv",names=['uid']).sort_values(by=['uid'])
all_data = train.append(test)
all_data = all_data.sort_values(by=['uid']).reset_index(drop=True)
print(all_data.shape)

from tqdm import tqdm

from collections import Counter

def Gini(pr):
    cate=Counter(pr)
    gini=1
    total=len(pr)
    for i in cate:
        gini=gini-(cate[i]/total)**2
    return gini

import math
# Calc Entropy
def entropy(pr):
    cate = Counter(pr)
    log2 = math.log2
    total = len(pr)
    ent = 0
    for i in cate:
        p = float(cate[i] / total)
        if p==0:
            ent=0
            continue
        ent = ent - p * (log2(p))
    return ent

def get_small(x,a,b):
    return int((x>a) & (x<b))

def get_feature_flatten(df):
    
    fea = []
    # Count Encoder
    t1 = time.time()
    df['appid_count'] = df.groupby(['appid'])['uid'].transform('count')
    fea.append(df[['uid','appid_count']].groupby(['uid'])['appid_count'].agg({'mean','std','min','max','median'}))
    fea.append(df[['uid','appid']].groupby(['uid'])['appid'].apply(Gini))
    fea.append(df[['uid','appid']].groupby(['uid'])['appid'].apply(entropy))
    df['appid_count_0_1e3'] = df['appid_count'].map(lambda x:get_small(x,0,1e3))
    df['appid_count_1e3_1e4'] = df['appid_count'].map(lambda x:get_small(x,1e3,1e4))
    df['appid_count_1e4_2e5'] = df['appid_count'].map(lambda x:get_small(x,1e4,2e5))
    tmp = df.groupby(['uid']).agg({'appid_count_0_1e3' : ['mean','sum','std'],
                          'appid_count_1e3_1e4' : ['mean','sum','std'],
                          'appid_count_1e4_2e5' : ['mean','sum','std']
                          })

    tmp.columns = ['_'.join(col).strip() for col in tmp.columns.values]
    fea.append(tmp)
    print("Count Active Finish... :",time.time()-t1)
    
    return fea


uid0 = all_data[['uid']]
fea0 = get_feature_flatten(usage)
for i in tqdm(fea0):
    uid0 = uid0.merge(i,how='left',on='uid')

uid0 = uid0.set_index('uid').add_prefix("usage_").reset_index()

uid1 = all_data[['uid']]
fea1 = get_feature_flatten(active)

for i in tqdm(fea1):
    uid1 = uid1.merge(i,how='left',on='uid')
uid1 = uid1.set_index('uid').add_prefix("active_").reset_index()

uid0.to_pickle("../pickle/usage_flatten_stat.pickle")
uid1.to_pickle("../pickle/active_flatten_stat.pickle")

uid1
