==> snake/9. LGB COUNT & TFIDF Model.ipynb <==
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import lightgbm as lgb
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder,OneHotEncoder
from sklearn.model_selection import KFold,StratifiedKFold
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score,roc_auc_score,f1_score,recall_score
from sklearn.feature_extraction.text import CountVectorizer,TfidfVectorizer,HashingVectorizer
from sklearn.decomposition import TruncatedSVD,SparsePCA
import gc
import time
import os
import sys
import warnings
warnings.filterwarnings('ignore')

train = pd.read_csv("../data/age_train.csv",names=['uid','age_group']).sort_values(by=['uid'])
test = pd.read_csv("../data/age_test.csv",names=['uid']).sort_values(by=['uid'])
info = pd.read_csv("../data/app_info.csv",names=['appid','category'])
active = pd.read_pickle("../pickle/user_app_active.pickle")
# usage = pd.read_pickle("../input2/user_app_usage.pickle")#,names=['uid','appid','duration','times','use_date'],parse_dates=['use_date'])
user_basic_info = pd.read_csv("../data/user_basic_info.csv",names=['uid','gender','city','prodname','ramcapacity','ramleftration','romcapacity','romleftration','color','fontsize','ct','carrier','os']).sort_values(by=['uid'])
behavior_info = pd.read_csv("../data/user_behavior_info.csv",names=['uid','boottimes','a','b','c','d','e','f','g']).sort_values(by=['uid'])
# (train.shape,test.shape),(info.shape,active.shape,usage.shape,user_basic_info.shape,behavior_info.shape)

all_data = train.append(test).reset_index(drop=True)
all_data.head()

def get_category(x):
    col = []
    no_col = 0
    for i in x:
        try:
            col.append(hash_dict[i])
        except:
            no_col+=1
    return col,no_col

hash_dict = dict(info.values)
active['category'] = active['appid'].map(lambda x:get_category(x))
active['category_nan']  = active['category'].map(lambda x:x[1])
active['category']  = active['category'].map(lambda x:x[0])
active['category_len'] = active['category'].map(lambda x:len(x))
active['category_nunique'] = active['category'].map(lambda x:len(set(x)))
active['category_ratio'] = active['category_nunique']/active['category_len']
del active['category']

all_data = all_data.merge(user_basic_info,how='left',on=['uid'])
all_data = all_data.merge(behavior_info,how='left',on=['uid'])
all_data = all_data.merge(active,how='left',on=['uid'])
feature_name = [i for i in all_data.columns if i not in ['uid','age_group']]
all_data.shape,len(all_data)

all_data['city_count_user'] = all_data.groupby(['city'])['uid'].transform('count')
all_data['romleftration_count_user'] = all_data.groupby(['romleftration'])['uid'].transform('count')
all_data['prodname_count_user'] = all_data.groupby(['prodname'])['uid'].transform('count')
all_data['color_count_user'] = all_data.groupby(['color'])['uid'].transform('count')
all_data['ct_count_user'] = all_data.groupby(['ct'])['uid'].transform('count')
all_data['carrier_count_user'] = all_data.groupby(['carrier'])['uid'].transform('count')

all_data['city_nunique_user'] = all_data.groupby(['city'])['uid'].transform('nunique')
all_data['romleftration_nunique_user'] = all_data.groupby(['romleftration'])['uid'].transform('nunique')
all_data['prodname_nunique_user'] = all_data.groupby(['prodname'])['uid'].transform('nunique')
all_data['ct_nunique_user'] = all_data.groupby(['ct'])['uid'].transform('nunique')
all_data['carrier_nunique_user'] = all_data.groupby(['carrier'])['uid'].transform('nunique')

all_data.head()

feature_name = [i for i in all_data.columns if i not in ['uid','age_group']]
cat_col = [col for col in all_data.columns if all_data[col].dtype == np.object]
num_col = [i for i in feature_name  if i not in cat_col]

label_name = ['age_group']
from tqdm import tqdm
from scipy import sparse
vector_feature = ['appid']
onehot_feature =  [i for i in cat_col if i not in vector_feature]

tr = None
train_ix = list(range(train.shape[0]))
test_ix = list(range(train.shape[0],all_data.shape[0]))

print('onehot...')
enc = OneHotEncoder(handle_unknown='ignore')
for feature in tqdm(onehot_feature):
    lbl = LabelEncoder()
    all_data[feature] = lbl.fit_transform(all_data[feature].astype('str').fillna('0').values.reshape(-1, 1))

from scipy import sparse

c1 = sparse.load_npz("../vector/Sparse_Matrix/active_count.npz")
c2 = sparse.load_npz("../vector/Sparse_Matrix/active_tfidf.npz")

svd = TruncatedSVD(n_components=100,n_iter=10,random_state=2019)
c1_svd = svd.fit_transform(c1)
c2_svd = svd.fit_transform(c2)

print('cv...')
cv = TfidfVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=2)
for feature in tqdm(vector_feature):
    cv.fit(all_data[feature].astype('str'))
    train_a = cv.transform(all_data[feature].astype('str').fillna("##").loc[train_ix])
    test_a = cv.transform(all_data[feature].astype('str').fillna("##").loc[test_ix])
    if tr is None:
        tr = train_a
        te = test_a
    else:
        tr = sparse.hstack((tr, train_a), 'csr')
        te = sparse.hstack((te, test_a), 'csr')    
print(tr.shape,te.shape)
    
cv = CountVectorizer(analyzer='word',token_pattern=u"(?u)\\b\\w+\\b",min_df=2)
for feature in tqdm(vector_feature):
    cv.fit(all_data[feature].astype('str'))
    train_a = cv.transform(all_data[feature].astype('str').fillna("##").loc[train_ix])
    test_a = cv.transform(all_data[feature].astype('str').fillna("##").loc[test_ix])
    tr = sparse.hstack((tr, train_a), 'csr')
    te = sparse.hstack((te, test_a), 'csr')

print(tr.shape,te.shape)

feature_name = [i for i in all_data.columns if i not in ['uid','age_group']]
tr = sparse.hstack((tr,all_data.loc[train_ix][num_col]),'csr')
te = sparse.hstack((te,all_data.loc[test_ix][num_col]),'csr')
print(tr.shape,te.shape)

import catboost as cbt
def acc_score(labels, preds):

#     preds = check_prob(preds.reshape(12,-1))
#     preds = np.argmax(preds, axis=0)
    preds = np.argmax(preds, axis=1)
    score = accuracy_score(y_true=preds, y_pred=labels)
    return 'acc_score', score, True

tr_index = ~all_data[label_name].isnull()
X_train = tr#all_data[tr_index][list(set(feature_name))].reset_index(drop=True)
y = train['age_group'].values - 1
X_test = te#all_data[~tr_index][list(set(feature_name))].reset_index(drop=True)
final_pred = []
cv_score = []
cv_model = []

skf = StratifiedKFold(n_splits=5, random_state=2019, shuffle=True)
for index, (train_index, test_index) in enumerate(skf.split(X_train, y)):
    print(index)
    lgb_model = lgb.LGBMClassifier(
        boosting_type="gbdt", num_leaves=128, reg_alpha=0.1, reg_lambda=1,
        max_depth=-1, n_estimators=3000, objective='multiclass',num_class=6,
        subsample=0.5, colsample_bytree=0.5, subsample_freq=1,min_child_samples=20,
        learning_rate=0.1, random_state=2019 + index, n_jobs=50, metric="multi_error", importance_type='gain'
    )
#     train_x, test_x, train_y, test_y = X_train.iloc[train_index], X_train.iloc[test_index], y.iloc[train_index], y.iloc[test_index]
    train_x, test_x, train_y, test_y = X_train[train_index], X_train[test_index], y[train_index], y[test_index]
    eval_set = [(test_x, test_y)]
    lgb_model.fit(train_x, train_y, eval_set=eval_set,verbose=10,early_stopping_rounds=100)
    # cbt_model = cbt.CatBoostClassifier(n_estimators=3000,learning_rate=0.1,max_depth=7,colsample_bytree=0.5,objective='multi_class',verbose=10,early_stopping_rounds=100,task_type='GPU',eval_metric='Accuracy')
    # cbt_model.fit(train_x, train_y,eval_set=[test_x,test_y])
    cv_model.append([lgb_model])
    y_test = lgb_model.predict(X_test)
    y_val = lgb_model.predict_proba(test_x)
    cv_score.append(acc_score(test_y,y_val)[1])
    if index == 0:
        final_pred = np.array(y_test).reshape(-1, 1)
    else:
        final_pred = np.hstack((final_pred, np.array(y_test).reshape(-1, 1)))

print("LGBM : ",np.mean(cv_score))

# skf = StratifiedKFold(n_splits=5, random_state=2019, shuffle=True)
cv_pred_lgb = np.zeros((X_train.shape[0],6))
test_pred_lgb = np.zeros((X_test.shape[0],6))

for index, (train_index, test_index) in enumerate(skf.split(X_train, y)):
    print(index)
    train_x, test_x, train_y, test_y = X_train[train_index], X_train[test_index], y[train_index], y[test_index]
    y_val = cv_model[index][0].predict_proba(test_x)
    cv_pred_lgb[test_index] = y_val
    print(y_val[:5])
    test_pred_lgb += cv_model[index][0].predict_proba(X_test) / 5

%%time
k = cv_model[0][0].predict_proba(test_x)

oof = pd.DataFrame(cv_pred_lgb)
oof['uid'] = train['uid'].values

pred = pd.DataFrame(test_pred_lgb)
pred['uid'] = test['uid'].values
